+++ CHANGELOG.md
@@ -53,3 +53,6 @@ Initial version
 ## Version 0.2.4 (xx.xx.2020)
 * visualization
   * added absolute time to hover info of gpio plots
+* xmlValidate now returns value (string) if execution fails
+* CLI
+  * return exit status 1 if command failed